Add guide to explain how to opt-in to Hermes.
Unfortunately there is no good way to create a new docs page for the current version in Docusaurus. I decided to wipe away the 0.60-RC version and re-create the versioned docs at 0.60, including Hermes. There haven't been any major changes to the docs since the release so there shouldn't be any real gaps between the docs about master and 0.60. If there are, we'll fix them forward.